Part 4: 4.21 - 4.22

+++ controllers/blogs.js
@@ -1,8 +1,5 @@
 const Blog = require('../models/blog')
-const User = require("../models/user");
 const blogRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
-
 
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
@@ -11,11 +8,10 @@ blogRouter.get('/', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({error: 'token missing or invalid'})
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({error: 'invalid token'})
   }
-  const user = await User.findById(decodedToken.id)
 
   body.user = user._id
   const blog = new Blog(body)
@@ -26,6 +22,19 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+  if (!user) {
+    return response.status(401).json({error: 'invalid token'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({error: 'blog not found'})
+  }
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({error: 'blog can only be deleted by creator'})
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
